requirements.txt
mbed-cli==1.0.0
pyserial>=3.2.1
